Fix more Linter error
